Move some function torecording_tools
