+++ src/spikeinterface/core/baserecording.py
@@ -12,9 +12,12 @@ from .core_tools import (
     check_json,
     convert_bytes_to_str,
     convert_seconds_to_str,
+)
+from .recording_tools import (
     write_binary_recording,
     write_memory_recording,
     write_traces_to_zarr,
+
 )
 from .job_tools import split_job_kwargs
